Add Crud

+++ src/main/java/Project/Security/user/User.java
@@ -1,5 +1,6 @@
 package Project.Security.user;
 
+import Project.Security.Role.Role;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -21,12 +22,12 @@ import java.util.List;
 public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String firstname;
     private String lastname;
     private String email;
     private String password;
-    private int age;
+    private Integer age;
     @Enumerated(EnumType.STRING)
     private Role role;
     @Override
